Move relative path generator to utils and make name more generic
